+++ src/lib/isIOS.js
@@ -1,3 +1,5 @@
+import getOS from './getOS';
+
 /**
  * 检查是不是iOS环境。
  * @returns {boolean} 真/假
@@ -12,7 +14,7 @@
  *
  */
 function isIOS() {
-  return /(iPhone|iPad|iPod|iOS)/i.test(navigator.userAgent);
+  return getOS().name === 'iOS';
 }
 
 export default isIOS;