+++ src/lib/isLinux.js
@@ -1,3 +1,5 @@
+import getOS from "./getOS";
+
 /**
  * 检查是不是Linux环境。
  * @returns {boolean} 真/假
@@ -12,7 +14,7 @@
  *
  */
 function isLinux() {
-  return navigator.platform.indexOf('Linux') > -1;
+  return getOS().name === 'Linux';
 }
 
 export default isLinux;